+++ tests/testthat/test.DirectAdjusted.R
@@ -365,7 +365,7 @@ test_that("absorbed_intercepts", {
   expect_false(noblocks_lmitt_fitted_not_absorbed@absorbed_intercepts)
 })
 
-test_that("absorbed_moderators", {
+test_that("@moderator slot", {
   data(simdata)
 
   blockeddes <- rct_design(z ~ block(bid) + cluster(cid1, cid2), data = simdata)
@@ -381,11 +381,11 @@ test_that("absorbed_moderators", {
   not_lmitt_fitted <- as.lmitt(lm(y ~ assigned(noblocksdes), data = simdata),
                                design = noblocksdes)
 
-  expect_equal(noblocks_lmitt_fittedsbgrp@absorbed_moderators, "force")
-  expect_equal(blocked_lmitt_fittedsbgrp@absorbed_moderators, "force")
-  expect_equal(lmitt_fitted_nosbgrp@absorbed_moderators, vector("character"))
-  expect_equal(blocked_lmitt_fitted_nosbgrp@absorbed_moderators, vector("character"))
-  expect_equal(not_lmitt_fitted@absorbed_moderators, vector("character"))
+  expect_equal(noblocks_lmitt_fittedsbgrp@moderator, "force")
+  expect_equal(blocked_lmitt_fittedsbgrp@moderator, "force")
+  expect_equal(lmitt_fitted_nosbgrp@moderator, vector("character"))
+  expect_equal(blocked_lmitt_fitted_nosbgrp@moderator, vector("character"))
+  expect_equal(not_lmitt_fitted@moderator, vector("character"))
 })
 
 test_that("estfun.DirectAdjusted requires a certain model class", {
@@ -464,7 +464,7 @@ test_that(paste("estfun.DirectAdjusted returns correct dimensions and alignment"
   mod2 <- lmitt(y ~ 1, data = shuffled_Q_data, design = des, offset = cov_adj(cmod, by = "uid"))
 
   expect_equal(dim(estfun(mod1)), c(nrow(simdata), 2))
-  expect_equal(estfun(mod1)[1:20,], estfun(mod2)[1:20,], tolerance = 1e8)
+  expect_equal(estfun(mod1), estfun(mod2))
 })
 
 test_that(paste("estfun.DirectAdjusted returns correct dimensions for partial",
@@ -715,10 +715,10 @@ test_that(paste(".align_and_extend_estfuns when exact alignment of C and Q is",
 
   expect_equal(dim(ef1$phi), c(nrow(simdata), 2))
   expect_equal(dim(ef1$psi), c(nrow(simdata), 2))
-  expect_true(all(ef1$phi[1:20,] == 0))
+  expect_true(all(ef1$phi[21:50,] == 0))
   expect_equal(ef1$phi, ef2$phi)
   expect_equal(ef1$psi, ef1$psi[sort(simdata$obs_id, index.return = TRUE)$ix,, drop = FALSE])
-  expect_true(all(ef1$psi[31:50,] == 0))
+  expect_true(all(ef1$psi[1:20,] == 0))
   expect_equal(ef1$psi, ef2$psi)
 })
 
@@ -862,15 +862,15 @@ test_that(paste(".align_and_extend_estfuns when exact alignment of C and Q isn't
   phi2_sorted <- lapply(split(ef2$phi, by_ix), sort)
   expect_true(all(sapply(unique(by_ix),
                          function(id) all.equal(phi1_sorted[[id]], phi2_sorted[[id]]))))
-  expect_true(all(ef1$phi[1:20,] == 0))
-  expect_true(all(ef2$phi[1:20,] == 0))
+  expect_true(all(ef1$phi[21:50,] == 0))
+  expect_true(all(ef2$phi[21:50,] == 0))
 
   psi1_sorted <- lapply(split(ef1$psi, by_ix), sort)
   psi2_sorted <- lapply(split(ef2$psi, by_ix), sort)
   expect_true(all(sapply(unique(by_ix),
                          function(id) all.equal(psi1_sorted[[id]], psi2_sorted[[id]]))))
-  expect_true(all(ef1$psi[31:50,] == 0))
-  expect_true(all(ef2$psi[31:50,] == 0))
+  expect_true(all(ef1$psi[1:20,] == 0))
+  expect_true(all(ef2$psi[1:20,] == 0))
 })
 
 test_that(".make_uoa_ids fails without cluster argument or DirectAdjusted model", {
@@ -974,6 +974,36 @@ test_that(".make_uoa_ids returns correct ID's for partial overlap of C and Q", {
   expect_equal(length(unique(ids)), length(unique(Q_uoas)) + 20)
 })
 
+test_that(paste(".make_uoa_ids returns correct ID's when cov_adj's 'by' argument",
+                "provides a different ordering"), {
+  data(simdata)
+  simdata$id <- sample(seq_len(nrow(simdata)))
+  
+  set.seed(300)
+  C_not_Q <- data.frame("y" = rnorm(20), "x" = rnorm(20), "cid1" = NA, "cid2" = NA,
+                        "id" = seq(51, 70))
+  cmod_data <- rbind(simdata[, colnames(C_not_Q)], C_not_Q)
+  
+  cmod <- lm(y ~ x, cmod_data)
+  des <- rct_design(z ~ uoa(cid1, cid2), simdata)
+  dmod <- lmitt(y ~ 1, data = simdata, design = des, offset = cov_adj(cmod, by = "id"))
+  
+  Q_uoas <- apply(simdata[, c("cid1", "cid2"), drop = FALSE], 1,
+                  function(...) paste(..., collapse = "_"))
+  
+  ids <- .make_uoa_ids(dmod)
+  
+  expect_true(is.factor(ids))
+  
+  expect_equal(length(ids), nrow(simdata) + 20)
+  
+  expect_true(all.equal(ids[1:nrow(simdata)],
+                        factor(Q_uoas)[sort(simdata$id, index.return = TRUE)$ix],
+                        check.attributes = FALSE))
+  
+  expect_equal(length(unique(ids)), length(unique(Q_uoas)) + 20)
+})
+
 test_that(paste(".order_samples fails without a DirectAdjusted object or",
                 "SandwichLayer offset"), {
   data(simdata)
@@ -1000,12 +1030,15 @@ test_that(".order_samples when the samples fully overlap", {
 
   expect_equal(length(out$Q_order), nrow(simdata))
   expect_equal(length(out$C_order), nrow(simdata))
+  expect_equal(length(out$Q_union_C_order), nrow(simdata))
 
   expect_equal(names(out$Q_order), as.character(seq_len(nrow(simdata))))
   expect_equal(names(out$C_order), as.character(seq_len(nrow(simdata))))
-
-  expect_equal(sum(is.na(out$Q_order)), 0)
-  expect_equal(sum(is.na(out$C_order)), 0)
+  expect_equal(names(out$Q_union_C_order), as.character(seq_len(nrow(simdata))))
+  
+  expect_true(all.equal(out$Q_order, seq_len(nrow(simdata)), check.attributes = FALSE))
+  expect_true(all.equal(out$C_order, seq_len(nrow(simdata)), check.attributes = FALSE))
+  expect_true(all.equal(out$Q_union_C_order, seq_len(nrow(simdata)), check.attributes = FALSE))
 })
 
 test_that(".order_samples when Q is a subset of C", {
@@ -1013,8 +1046,11 @@ test_that(".order_samples when Q is a subset of C", {
   data(simdata)
 
   simdata$uid <- seq_len(nrow(simdata))
-  cmod_data <- data.frame(x = rnorm(30), y = rnorm(30), cid1 = NA, cid2 = NA,
-                          uid = seq_len(30) + nrow(simdata))
+  cmod_data <- rbind(
+    simdata[, c("x", "y", "cid1", "cid2", "uid")],
+    data.frame(x = rnorm(30), y = rnorm(30), cid1 = NA, cid2 = NA,
+               uid = seq_len(30) + nrow(simdata))
+  )
   cmod <- lm(y ~ x, cmod_data)
   des <- rct_design(z ~ cluster(cid1, cid2), data = simdata)
   mod <- lmitt(y ~ 1, data = simdata, design = des, offset = cov_adj(cmod, by = "uid"))
@@ -1023,12 +1059,15 @@ test_that(".order_samples when Q is a subset of C", {
 
   expect_equal(length(out$Q_order), nrow(simdata))
   expect_equal(length(out$C_order), nrow(simdata) + 30)
+  expect_equal(length(out$Q_union_C_order), nrow(simdata) + 30)
 
   expect_equal(names(out$Q_order), as.character(seq_len(nrow(simdata))))
   expect_equal(names(out$C_order), as.character(seq_len(nrow(simdata) + 30)))
-
-  expect_equal(sum(is.na(out$Q_order)), 0)
-  expect_equal(sum(is.na(out$C_order)), nrow(simdata))
+  expect_equal(names(out$Q_union_C_order), as.character(seq_len(nrow(simdata) + 30)))
+  
+  expect_true(all.equal(out$Q_order, seq_len(nrow(simdata)), check.attributes = FALSE))
+  expect_true(all.equal(out$C_order, seq_len(nrow(simdata) + 30), check.attributes = FALSE))
+  expect_true(all.equal(out$Q_union_C_order, seq_len(nrow(simdata) + 30), check.attributes = FALSE))
 })
 
 
@@ -1045,13 +1084,16 @@ test_that(".order_samples when C is a subset of Q", {
   out <- .order_samples(mod)
 
   expect_equal(length(out$Q_order), nrow(simdata))
-  expect_equal(length(out$C_order), nrow(simdata))
+  expect_equal(length(out$C_order), 20)
+  expect_equal(length(out$Q_union_C_order), nrow(simdata))
 
   expect_equal(names(out$Q_order), as.character(seq_len(50)))
-  expect_equal(names(out$C_order), as.character(seq_len(50)))
+  expect_equal(names(out$C_order), as.character(seq_len(20)))
+  expect_equal(names(out$Q_union_C_order), as.character(seq_len(50)))
 
-  expect_equal(sum(is.na(out$Q_order)), 0)
-  expect_equal(sum(is.na(out$C_order)), 30)
+  expect_true(all.equal(out$Q_order, seq_len(nrow(simdata)), check.attributes = FALSE))
+  expect_true(all.equal(out$C_order, seq_len(20), check.attributes = FALSE))
+  expect_true(all.equal(out$Q_union_C_order, seq_len(nrow(simdata)), check.attributes = FALSE))
 })
 
 test_that(".order_samples when the samples do not overlap", {
@@ -1068,13 +1110,16 @@ test_that(".order_samples when the samples do not overlap", {
   out <- .order_samples(mod)
 
   expect_equal(length(out$Q_order), 30)
-  expect_equal(length(out$C_order), nrow(simdata))
+  expect_equal(length(out$C_order), 20)
+  expect_equal(length(out$Q_union_C_order), nrow(simdata))
 
   expect_equal(names(out$Q_order), as.character(seq(21, 50)))
-  expect_equal(names(out$C_order), as.character(c(seq(21, 50), seq_len(20))))
+  expect_equal(names(out$C_order), as.character(seq_len(20)))
+  expect_equal(names(out$Q_union_C_order), as.character(seq_len(nrow(simdata))))
 
-  expect_equal(sum(is.na(out$Q_order)), 0)
-  expect_equal(sum(is.na(out$C_order)), 30)
+  expect_true(all.equal(out$Q_order, seq_len(30), check.attributes = FALSE))
+  expect_true(all.equal(out$C_order, seq_len(20), check.attributes = FALSE))
+  expect_true(all.equal(out$Q_union_C_order, c(seq(31, 50), seq_len(30)), check.attributes = FALSE))
 })
 
 test_that(".order_samples when no `by` argument provided", {
@@ -1096,13 +1141,16 @@ test_that(".order_samples when no `by` argument provided", {
   names(C_uoas) <- NULL
 
   expect_equal(length(out$Q_order), 30)
-  expect_equal(length(out$C_order), nrow(simdata))
+  expect_equal(length(out$C_order), 20)
+  expect_equal(length(out$Q_union_C_order), nrow(simdata))
 
   expect_equal(names(out$Q_order), Q_uoas)
-  expect_equal(names(out$C_order), c(Q_uoas, C_uoas))
+  expect_equal(names(out$C_order), C_uoas)
+  expect_equal(names(out$Q_union_C_order), sort(c(Q_uoas, C_uoas)))
 
-  expect_equal(sum(is.na(out$Q_order)), 0)
-  expect_equal(sum(is.na(out$C_order)), 30)
+  expect_true(all.equal(out$Q_order, seq_len(30), check.attributes = FALSE))
+  expect_true(all.equal(out$C_order, seq_len(20), check.attributes = FALSE))
+  expect_true(all.equal(out$Q_union_C_order, c(seq(31, 50), seq_len(30)), check.attributes = FALSE))
 })
 
 test_that("sanitize_Q_ids fails with invalid cluster argument", {