+++ tests/testthat/test.lmitt.R
@@ -300,6 +300,69 @@ test_that("non-integer units of assignment", {
 
 })
 
+test_that("#137 ensure absorb is using the correct moderator", {
+  data(simdata)
+
+  mod1 <- lmitt(y ~ o, data = simdata,
+                design = z ~ cluster(cid1, cid2) + block(bid))
+  mod2 <- lmitt(y ~ o, data = simdata,
+                design = z ~ cluster(cid1, cid2) + block(bid),
+                absorb = TRUE)
+
+  o1 <- model.matrix(mod1)[, "o"]
+  o2 <- model.matrix(mod2)[, "o"]
+
+  # o1 and o2 should be different, since the latter should have been
+  # group-centered due to `absorb = TRUE`. Due to a bug discovered in #137, the
+  # un-centered version of the continuous moderator was being using previously,
+  # leading o1 and o2 to be identical (incorrectly).
+
+  expect_true(!(all.equal(o1, o2, check.attributes = FALSE) == TRUE))
+
+
+})
+
+test_that("#140 handling 0 weights", {
+
+  # A single 0 weight observation in a larger block - that observation's
+  # conbributiont to the estimating equation should be 0
+  data(simdata)
+  simdata$weight <- 1
+  simdata$weight[1] <- 0
+
+  des <- rct_design(z ~ uoa(cid1, cid2), data = simdata)
+  mod <- lmitt(y ~ x, data = simdata, design = des,
+               weights = simdata$weight)
+  ee <- propertee:::estfun.DirectAdjusted(mod)
+  expect_true(all(ee[1, ] == 0))
+
+  # A block with only a single non-zero weight should contribute nothing to the
+  # estimating equation
+  data(simdata)
+  simdata$weight <- 1
+  simdata$weight[simdata$bid == 1] <- 0
+  simdata$weight[simdata$bid == 1][1] <- 1
+
+  des <- rct_design(z ~ uoa(cid1, cid2) + block(bid), data = simdata)
+  mod <- lmitt(y ~ x, data = simdata, design = des,
+               absorb = TRUE, weights = simdata$weight)
+  ee <- propertee:::estfun.DirectAdjusted(mod)
+  ee[simdata$bid == 1, ]
+  #expect_true(all(ee[simdata$bid == 1, ] == 0))
+
+  data(simdata)
+  simdata$z[simdata$bid == 1] <- 1
+  simdata$weights <- 1
+  des <- rct_design(z ~ uoa(cid1, cid2) + block(bid), data = simdata)
+  # Block 1 has 0 variance in treatment
+  mod <- lmitt(y ~ x, data = simdata, design = des,
+               absorb = TRUE, weights = simdata$weight)
+  ee <- propertee:::estfun.DirectAdjusted(mod)
+  #expect_true(all(ee[simdata$bid == 1, ] == 0))
+
+})
+
+
 options(save_options)
 #### !!!!!!!!!!!NOTE!!!!!!!!!!!!!
 # This test below should NOT have `options()$propertee_message_on_unused_blocks`
@@ -326,3 +389,27 @@ test_that("Message if design has block info but isn't used in lmitt", {
 })
 
 ### READ COMMENT ABOUT LAST TEST!
+
+test_that(paste("absorb=TRUE doesn't drop rows when some strata have weights=0",
+                "(some strata only have treated clusters)"), {
+  data(simdata)
+  simdata[simdata$cid1 == 5 & simdata$cid2 == 2, "bid"] <- 4
+  
+  des <- rct_design(z ~ cluster(cid1, cid2) + block(bid), simdata)
+  absorb_mod <- lmitt(y ~ 1, design = des, data = simdata, weights = "ate", absorb = TRUE)
+  no_absorb_mod <- lmitt(y ~ 1, design = des, data = simdata, weights = "ate", absorb = FALSE)
+  
+  expect_equal(dim(model.matrix(absorb_mod)), dim(model.matrix(no_absorb_mod)))
+})
+
+test_that(paste("absorb=TRUE doesn't drop rows when some strata have weights=0",
+                "(some strata only have untreated clusters)"), {
+  data(simdata)
+  simdata[simdata$cid1 == 4 & simdata$cid2 == 2, "bid"] <- 4
+  
+  des <- rct_design(z ~ cluster(cid1, cid2) + block(bid), simdata)
+  absorb_mod <- lmitt(y ~ 1, design = des, data = simdata, weights = "ate", absorb = TRUE)
+  no_absorb_mod <- lmitt(y ~ 1, design = des, data = simdata, weights = "ate", absorb = FALSE)
+  
+  expect_equal(dim(model.matrix(absorb_mod)), dim(model.matrix(no_absorb_mod)))
+})